requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
openai==0.28.0
gradio==4.44.1

# aiohttp compatibility
aiohttp==3.9.3
typing_extensions==4.10.0
